refactor(config): move all AutoConfiguration classes to starter modules
