--- modules/i18n/dopamine-i18n/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-io.dopamine.i18n.config.I18nAutoConfiguration
