+++ modules/starter/dopamine-starter-mvc/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,2 @@
 io.dopamine.starter.mvc.config.ResponseAutoConfiguration
-io.dopamine.starter.mvc.config.TraceIdResolverConfig
+io.dopamine.starter.mvc.config.TraceIdAutoConfiguration
