--- modules/trace/dopamine-trace-mvc/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-io.dopamine.trace.mvc.config.TraceIdMvcAutoConfiguration
